Feature: Added more activities, fixed a lot of problems, finished working on this project
