+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,12 @@
         android:theme="@style/Theme.Calculator"
         tools:targetApi="31">
         <activity
+            android:name=".CalculatorsActivity"
+            android:exported="true"
+        />
+        <activity android:name=".AboutMeActivity"
+            android:exported="true"/>
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
