+++ app/src/main/java/com/example/calculator/Utils.kt
@@ -1,5 +1,8 @@
 package com.example.calculator
 
+import android.content.Context.MODE_PRIVATE
+import androidx.core.content.edit
+
 fun factorial(x: Int): Double{
     if (x<0){
         return Double.NaN
