+++ app/src/main/res/layout-land/about_me.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".AboutMeActivity"
     android:background="@color/background">
 
     <Button
