+++ app/src/main/res/layout-land/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".CalculatorsActivity"
     android:background="@color/background"
     >
