+++ app/src/main/res/layout-land/science_calculator.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".MainActivity">
+    tools:context=".CalculatorsActivity">
 
     <Button
         android:id="@+id/backToMenuButton"
