+++ app/src/main/res/layout/science_calculator.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".MainActivity">
+    tools:context=".CalculatorsActivity">
 
     <TextView
         android:id="@+id/calculationPanel"
