+++ app/src/main/res/values/themes.xml
@@ -1,7 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.Calculator" parent="Theme.AppCompat.Light.NoActionBar">
-    </style>
+    <style name="Base.Theme.Calculator" parent="Theme.AppCompat.Light.NoActionBar"></style>
 
     <style name="CustomActionButtonAC">
         <item name="android:background">@drawable/ac_button_selector</item>
@@ -51,6 +50,7 @@
         <item name="android:paddingRight">4dp</item>
     </style>
 
-    <style name="Theme.Calculator" parent="Base.Theme.Calculator" />
+    <style name="Theme.Calculator" parent="Base.Theme.Calculator"
+    />
 
 </resources>
